cirup-rs: - move revision struct into seperate file - fix handling of revision markers in filenames - fix visibility of modules
